+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## 0.2.0 - UNRELEASED
+## 0.2.0 - 9 Nov 2015
+
+Rename "categories" to "sets".
+Fix a bug where a dependencies.edn file more than one directory above
+the project root directory caused an exception.
 
 ## 0.1.0 - 6 Nov 2015